+++ scripts/lib/utilities/dialogUtils.js
@@ -156,7 +156,7 @@ async function confirm(title, content, {userId = game.user.id, buttons = 'yesNo'
 async function confirmUseItem(item, {userId = game.user.id, buttons = 'yesNo'} = {}) {
     return await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: item.name}), {userId, buttons});
 }
-async function selectDocumentDialog(title, content, documents, {displayTooltips = false, sortAlphabetical = false, sortCR = false, userId = game.user.id, addNoneDocument = false, showCR = false, showSpellLevel = false} = {}) {
+async function selectDocumentDialog(title, content, documents, {displayTooltips = false, sortAlphabetical = false, sortCR = false, userId = game.user.id, addNoneDocument = false, showCR = false, showSpellLevel = false, displayReference = false} = {}) {
     if (sortAlphabetical) {
         documents = documents.sort((a, b) => {
             return a.name.localeCompare(b.name, 'en', {'sensitivity': 'base'});
@@ -173,7 +173,8 @@ async function selectDocumentDialog(title, content, documents, {displayTooltips 
         name: isCompendiumDoc ? (i.uuid ?? i.actor?.uuid) : (i.id ?? i.actor?.id),
         options: {
             image: i.img + (i.system?.details?.cr != undefined ? ` (CR ${genericUtils.decimalToFraction(i.system?.details?.cr)})` : ``),
-            tooltip: displayTooltips ? i.system.description.value.replace(/<[^>]*>?/gm, '') : undefined
+            tooltip: displayTooltips ? i.system.description.value.replace(/<[^>]*>?/gm, '') : undefined,
+            reference: (displayReference && i.reference) ? i.reference : undefined
         }
     }));
     if (addNoneDocument) {