+++ scripts/macros/2024/monsterFeatures/brownBear/brownBearClaw.js
@@ -1,0 +1,8 @@
+export let brownBearClaw = {
+    name: 'Claw',
+    version: '1.3.132',
+    rules: 'modern',
+    monsters: [
+        'Brown Bear'
+    ]
+};