+++ scripts/macros/2024/monsterFeatures/mimic/mimicBite.js
@@ -1,0 +1,8 @@
+export let mimicBite = {
+    name: 'Bite',
+    version: '1.3.132',
+    rules: 'modern',
+    monsters: [
+        'Mimic'
+    ]
+};