+++ scripts/macros/2024/monsterFeatures/vampire/vampireCharm.js
@@ -1,6 +1,8 @@
 export let vampireCharm = {
     name: 'Charm',
-    monster: 'Vampire',
+    monsters: [
+        'Vampire'
+    ],
     version: '1.3.126',
     rules: 'modern'
 };