+++ scripts/macros/2024/monsterFeatures/vampire/vampireShapeShift.js
@@ -73,7 +73,9 @@ async function revert({trigger, workflow}) {
 }
 export let vampireShapeShift = {
     name: 'Shape-Shift',
-    monster: 'Vampire',
+    monsters: [
+        'Vampire'
+    ],
     version: '1.3.126',
     rules: 'modern',
     midi: {